backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[argon2]
google-generativeai
email-validator
